autentification part 1
admin almost done

+++ src/main/java/stasaaleksadavid/isabackend/controller/UserController.java
@@ -24,20 +24,31 @@ public class UserController {
 
     //get all
 
-    @GetMapping("/users")
-    public List<User> getAllUsers(){return userRepository.findAll();}
+    @GetMapping("/users/{type}")
+    public List<User> getAllUsers(@PathVariable String type){
+        if(type.equals("admin") || type.equals("main_admin")) {
+            return userRepository.findAll();
+        }
+        else{return null;}
+    }
 
     //create
-    @PostMapping("/users")
-    public  User createUser(@RequestBody User user){
+    @PostMapping("/users/{type}")
+    public  User createUser(@PathVariable String type,@RequestBody User user){
+        if(type.equals("admin") || type.equals("main_admin")) {
         return userRepository.save(user);
+        }
+        else{return null;}
     }
 
     //get by id
-    @GetMapping("/users/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable Long id){
+    @GetMapping("/users/{type}/{id}")
+    public ResponseEntity<User> getUserById(@PathVariable String type,@PathVariable Long id){
+        if(type.equals("admin") || type.equals("main_admin")) {
         User user = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User does not exist with id:"+ id));
         return ResponseEntity.ok(user);
+        }
+        else{return null;}
     }
     //update
     @PutMapping("/users/{type}/{id}")
@@ -45,9 +56,6 @@ public class UserController {
 
         if(type.equals("ship_owner") || type.equals("admin")|| type.equals("cottage_owner")|| type.equals("fishing_instructor")|| type.equals("main_admin")){
             User user = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User does not exist with id:" + id));
-
-
-
 
                 user.setPassword(userDetails.getPassword());
                 user.setFirstName(userDetails.getFirstName());
@@ -65,15 +73,17 @@ public class UserController {
             else{return null;}
     }
     //delete
-    @DeleteMapping("/users/{id}")
-    public Map<String, Boolean> deleteUser(@PathVariable Long id){
+    @DeleteMapping("/users/{type}/{id}")
+    public Map<String, Boolean> deleteUser(@PathVariable Long id,@PathVariable String type) {
+        if ( type.equals("admin") ||  type.equals("main_admin")) {
+            User user = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User does not exist with id:" + id));
 
-        User user = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User does not exist with id:"+ id));
-
-       userRepository.delete(user);
-        Map<String, Boolean> response = new HashMap<>();
-        response.put("deleted", Boolean.TRUE);
-        return (Map<String, Boolean>) ResponseEntity.ok(response);
+            userRepository.delete(user);
+            Map<String, Boolean> response = new HashMap<>();
+            response.put("deleted", Boolean.TRUE);
+            return (Map<String, Boolean>) ResponseEntity.ok(response);
+        }
+        else{return null;}
     }
 
     @GetMapping("/users/{email}/{password}")
